refactor: move authentication to /features and /hooks
