+++ gmt-web-app/src/App.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import Authentication from './components/Authentication';
+import Authentication from './features/auth/Authentication';
 
 const App: React.FC = () => {
   return (
